Per user Ansible hosts file. Option to leave power on of new OSM host.

+++ orchestrator/env/setup_from_btrfs.sh
@@ -2,6 +2,6 @@
 
 . setup_common.sh
 
-ansible-playbook -e "target=$vm_ip osm_host_name=$OSMHOST osm_domain=$OSM_DOMAIN fake_osm_host=True" -i /tmp/hosts osmhost_from_btrfs.yaml
+ansible-playbook -e "target=$vm_ip osm_host_name=$OSMHOST osm_domain=$OSM_DOMAIN fake_osm_host=True" -i "$ANSIBLE_HOSTS" osmhost_from_btrfs.yaml
 
-ssh root@$vm_ip "poweroff"
+[ -n "$POWER_ON" ] || ssh root@$vm_ip "poweroff"
